Refactor code before cucumber version 6 support
